Issue #1: Collaboration Rules Needed
Creates CONTRIBUTING.md to explain the rules and workflow for code changes.
Adds a "Creating Issues" section with a link and instructions.
Adds a "Committing Code" section that explains the Issue requirement and commit messages rules.
Adds a "Submitting Pull Requests" section to explain the git workflow, forks, and pull-requests.

Adds a Contributing section to README.md that links to the new CONTRIBUTING.md.