Issue #6: Added unit tests on POJO's and simple Post Example for Meds.
Created a simple MedWebService that has a method that can take in a
new medicine or existing medicine, and return a modified version as a
proof of concept for how API Gateway and Lambdas handle POST/PUT
methods with RequestBodies.

Also introduced apache-commons-3 for reflection methods for equals,
hashcode, and toString.

Also added 100% coverage tests for added POJOs to ensure all setters
that shouldn't taken in null don't and that all
equals/hashcode/toString methods return what is expected.