Issue #5: Adding DB support for Meds
This is a large commit that sets up the basic workflow for the server
side work.

First, created a ClientRequestObject.  The purpose  of this is to
handle the conversion in the API Gateway from an HTTP request into an
object with the necessary information from the request.  It contains
an accessToken, which gets deserialized into an AuthUser, a Map of
queryString objects from the URL, and the body, which is either an id
stored in the url or the request body passed into a PUT/POST request.

Next, updated the MySQLConnection to create connections to the new
hosted DB at Google CloudDB.

Created an entire DB populated workflow for Meds, allowing a user to
save a Med for a particular Amigo and to get meds for a particular
Amigo.