Issue #8: Added Auth0 validation
The Auth0 http requests passes back to the server an id token instead
of an access token.  Using the id token, we need to call into Jay's
auth0 setup to grab the access token information for google's stuff.
So I added an additional layer fixing that up.

Also fixed up the GET call on /meds to return the meds for the logged
in user's amigo.  This was our testing url.