Issue #9: Added CRUD operations for Amigos.
Added all support for the specific CRUD operations for users, as well
as adding a way to get the current user.  This also includes changes
that'll update the Picture url and name for every time the user logs
into the application, which will allow other user's to see each
other's profile pictures.