No issue attached: Added caching strategy for Auth Users
To improve performance, I've made drastic changes to how we handle id
tokens from Auth0.  Instead of re-authenticating with Google, I
instead grab the AuthUser's googleRef from the Auth0 call and get the
AuthUser from the DB that way.  Also, I created a new Table in the DB
for storing Tokens, the date added, and the number of seconds the
token is valid for.  If the token becomes invalid, it deletes the row
from that table, forcing a re-auth validation from auth0.  This
improves performance drastically across all lambda handlers.